+++ src/ctests/sdsc2.c
@@ -148,7 +148,7 @@ int main(int argc, char **argv)
       if (!TESTS_QUIET) {
          printf("\n(calculated independent of PAPI)\n");
          printf("\tOperations= %.1f Mflop", y * 1e-6);
-         printf("\t(%g Mflop/s)\n\n", ((float) y / (t2 - t1)));
+         printf("\t(%g Mflop/s)\n\n", (y / (double)(t2 - t1)));
          printf("PAPI measurements:\n");
       }
       for (j = 0; j < nevents; j++) {
